refactor(hu-128): unificar DuracionEnHorasDecimales a decimal en FranjaTemporal
Cambia el tipo de retorno de FranjaTemporal.DuracionEnHorasDecimales()
de double a decimal para alinearlo con la convencion del proyecto de usar
decimal en valores fraccionarios y evitar errores de IEEE 754 en calculos
de horas. Ningun consumidor de produccion dependia del tipo double.

+++ src/Bitakora.ControlAsistencia.Contracts/Programacion/ValueObjects/FranjaTemporal.cs
@@ -61,7 +61,7 @@ public abstract class FranjaTemporal
     public int DuracionEnMinutos() => MinutosAbsolutoFin - MinutosAbsolutoInicio;
 
     // CA-12: conversor a horas decimales
-    public double DuracionEnHorasDecimales() => DuracionEnMinutos() / (double)MinutosPorHora;
+    public decimal DuracionEnHorasDecimales() => DuracionEnMinutos() / (decimal)MinutosPorHora;
 
     // CA-20: formato legible - "(06:00-12:00)" o "(22:00-06:00+1)" con offset
     public abstract override string ToString();
